smart-control-main/requirements.txt
opencv-python
mediapipe
pyautogui
pillow
pystray
numpy